refactor: s3 업로드 타입에 인증샷 미션 추가 및 Facade 레이어 제거
