+++ src/main/java/com/example/live_backend/global/storage/controller/StorageController.java
@@ -1,14 +1,13 @@
 package com.example.live_backend.global.storage.controller;
 
 import com.example.live_backend.global.error.response.ResponseHandler;
-import com.example.live_backend.global.security.PrincipalDetails;
-import com.example.live_backend.global.storage.UploadFacade;
+import com.example.live_backend.global.security.annotation.AuthenticatedApi;
+import com.example.live_backend.global.storage.StorageService;
+import com.example.live_backend.global.storage.controller.docs.StorageControllerDocs;
 import com.example.live_backend.global.storage.dto.PresignedUrlRequestDto;
 import com.example.live_backend.global.storage.dto.PresignedUrlResponseDto;
-import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -19,17 +18,18 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 @Tag(name = "Storage", description = "파일 스토리지 API")
 
-public class StorageController {
+public class StorageController implements StorageControllerDocs {
 
-    private final UploadFacade uploadFacade;
+    private final StorageService storageService;
 
+    @Override
     @PostMapping("/presigned-url")
-    @Operation(summary = "Presigned URL 발급", description = "파일 업로드를 위한 Presigned URL를 발급합니다.")
+    @AuthenticatedApi(reason = "Presigned Url 발급 요청은 로그인한 사용자만 가능합니다.")
     public ResponseHandler<PresignedUrlResponseDto> generatePresignedUrl(
-            @RequestBody PresignedUrlRequestDto request,
-            @AuthenticationPrincipal PrincipalDetails userDetails) {
+            @RequestBody PresignedUrlRequestDto request
+    ) {
 
-        PresignedUrlResponseDto response = uploadFacade.generateUrlAndSave(request, userDetails);
+        PresignedUrlResponseDto response = storageService.generatePresignedUploadUrl(request);
 
         return ResponseHandler.success(response);
     }
